made sure that we detected the os platform the right way
now we are borrowing pyside from Shotgun installation if it does not exists. We are looking in the standard locations for it, but SHOTGUN_DESKTOP_INSTALL_PATH can be set to the directory where the Shotgun executable is located.

fixed the locations where the app can be found. CLARISSE_BIN_DIR can be defined to point to the directory where clarisse binary is located if required.

tweak to take into consideration minro versions of teh software, ie 4.0b